ai_assistant/requirements.txt
anthropic>=0.18.1
pyautogen>=0.2.0
python-dotenv>=1.0.0